+++ src/common/utxobased/info/ravencoin.ts
@@ -60,7 +60,7 @@ export const coinInfo: CoinInfo = {
   segwit: false,
   coinType: 175,
   prefixes: {
-    messagePrefix: ['\x18Bitcoin Signed Message:\n'],
+    messagePrefix: ['\x16Raven Signed Message:\n'],
     wif: [0x80],
     legacyXPriv: [0x0488ade4],
     legacyXPub: [0x0488b21e],