+++ src/common/utxobased/info/smartcash.ts
@@ -71,7 +71,7 @@ export const coinInfo: CoinInfo = {
   bs58EncodeFunc: base58smart.encode,
   wifEncodeFunc: wifsmart.encode,
   prefixes: {
-    messagePrefix: ['\x18Bitcoin Signed Message:\n'],
+    messagePrefix: ['\x1aSmartCash Signed Message:\n'],
     wif: [0xbf],
     legacyXPriv: [0x0488ade4],
     legacyXPub: [0x0488b21e],